requirements.txt
alembic==1.16.4
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
Django==5.2.5
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pdf2image==1.17.0
pillow==11.3.0
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.14.1
Werkzeug==3.1.3
yt-dlp==2025.8.11
